Make `build_command` a method
